requirements.txt
request
rembg
streamlit